[wrench_est][Plugin]: move i_term.h to wrench_est/
