+++ aerial_robot_control/include/aerial_robot_control/nmpc/tilt_qd_servo_thrust_dist_mdl/nmpc_controller_i_term.h
@@ -6,7 +6,7 @@
 #define TILT_QD_SERVO_THRUST_NMPC_W_ITERM_CONTROLLER_H
 
 #include "nmpc_controller.h"
-#include "aerial_robot_control/nmpc/i_term.h"
+#include "aerial_robot_control/wrench_est/i_term.h"
 
 #include "geometry_msgs/WrenchStamped.h"
